+++ variables.tf
@@ -30,6 +30,18 @@ variable "user_data" {
   default     = ""
 }
 
+variable "iam_instance_profile_name" {
+  type        = string
+  description = "(Optional, Forces new resource) The profile's name. If omitted, Terraform will assign a random, unique name."
+  default     = ""
+}
+
+variable "iam_role_name" {
+  type = string
+  description = "(Optional) The role name to include in the profile"
+  default     = ""
+}
+
 # https://github.com/hashicorp/terraform/issues/24188
 # lifecycle: Support for dynamic blocks and meta-arguments still open
 // variable "lifecycle_ignore_changes" {