+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ENHANCEMENTS:
 
 * Added optional user_data configuration [#1](https://github.com/zoitech/terraform-aws-ec2/issues/1)
+* Added optional iam_instance_profile configuration [#2](https://github.com/zoitech/terraform-aws-ec2/issues/2)
 
 ## 1.0.0